+++ src/test/php/text/parse/unittest/rules/MatchTest.class.php
@@ -1,18 +1,18 @@
 <?php namespace text\parse\unittest\rules;
 
 use text\parse\Tokenized;
-use text\parse\rules\Match;
+use text\parse\rules\Matches;
 use unittest\{Test, Values};
 
 class MatchTest extends \unittest\TestCase {
 
   #[Test, Values([['1', 1], ['-1', -1], ['1.5', 1.5], ['-1.5', -1.5], ['true', true]])]
   public function constants_and_numbers($input, $outcome) {
-    $rule= new Match([
+    $rule= new Matches([
       T_STRING  => function($values) { return constant($values[0]); },
       T_DNUMBER => function($values) { return (double)$values[0]; },
       T_LNUMBER => function($values) { return (int)$values[0]; },
-      '-'       => new Match([
+      '-'       => new Matches([
         T_DNUMBER => function($values) { return -(double)$values[0]; },
         T_LNUMBER => function($values) { return -(int)$values[0]; },
       ])
@@ -22,7 +22,7 @@ class MatchTest extends \unittest\TestCase {
 
   #[Test]
   public function mismatched_numbers() {
-    $rule= new Match([
+    $rule= new Matches([
       T_DNUMBER => function($values) { /* Intentionally empty */ },
       T_LNUMBER => function($values) { /* Intentionally empty */ }
     ]);